+++ modules/juce_core/zip/juce_GZIPCompressorOutputStream.cpp
@@ -46,7 +46,7 @@ public:
             zlibNamespace::deflateEnd (&stream);
     }
 
-    bool write (const uint8* data, int dataSize, OutputStream& out)
+    bool write (const uint8* data, unsigned int dataSize, OutputStream& out)
     {
         // When you call flush() on a gzip stream, the stream is closed, and you can
         // no longer continue to write data to it!
@@ -62,7 +62,7 @@ public:
     void finish (OutputStream& out)
     {
         const uint8* data = nullptr;
-        int dataSize = 0;
+        unsigned int dataSize = 0;
 
         while (! finished)
             doNextBlock (data, dataSize, out, Z_FINISH);
@@ -76,7 +76,7 @@ private:
     bool isFirstDeflate, streamIsValid, finished;
     zlibNamespace::Bytef buffer[32768];
 
-    bool doNextBlock (const uint8*& data, int& dataSize, OutputStream& out, const int flushMode)
+    bool doNextBlock (const uint8*& data, unsigned int& dataSize, OutputStream& out, const int flushMode)
     {
         using namespace zlibNamespace;
         if (streamIsValid)
@@ -98,7 +98,7 @@ private:
                 case Z_OK:
                 {
                     data += dataSize - stream.avail_in;
-                    dataSize = (int) stream.avail_in;
+                    dataSize = stream.avail_in;
                     const int bytesDone = ((int) sizeof (buffer)) - (int) stream.avail_out;
                     return bytesDone <= 0 || out.write (buffer, bytesDone);
                 }
@@ -140,7 +140,8 @@ bool GZIPCompressorOutputStream::write (const void* destBuffer, int howMany)
 {
     jassert (destBuffer != nullptr && howMany >= 0);
 
-    return helper->write (static_cast <const uint8*> (destBuffer), howMany, *destStream);
+    return helper->write (static_cast <const uint8*> (destBuffer),
+                          (unsigned int) howMany, *destStream);
 }
 
 int64 GZIPCompressorOutputStream::getPosition()
@@ -176,7 +177,7 @@ public:
 
                 for (int j = rng.nextInt (100); --j >= 0;)
                 {
-                    MemoryBlock data (rng.nextInt (2000) + 1);
+                    MemoryBlock data ((unsigned int) (rng.nextInt (2000) + 1));
 
                     for (int k = (int) data.getSize(); --k >= 0;)
                         data[k] = (char) rng.nextInt (255);
